+++ tests/beman/inplace_vector/template.test.cpp
@@ -1,0 +1,5 @@
+#include <gtest/gtest.h>
+
+#include "gtest_setup.hpp"
+
+namespace {};
